# http/openapi_types.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen/v2 version v2.2.0 DO NOT EDIT.
package http

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Defines values for TaskStatus.
const (
	TASKCOMPLETED   TaskStatus = 1
	TASKUNCOMPLETED TaskStatus = 0
)

// Error defines model for Error.
type Error struct {
	// Message Error message
	Message string `json:"message"`
}

// Task defines model for Task.
type Task struct {
	CreatedAt time.Time          `json:"createdAt"`
	Id        openapi_types.UUID `json:"id"`

	// Name task name
	Name      string     `json:"name"`
	Status    TaskStatus `json:"status"`
	UpdatedAt time.Time  `json:"updatedAt"`
}

// TaskStatus defines model for Task.Status.
type TaskStatus int

// PostTaskJSONRequestBody defines body for PostTask for application/json ContentType.
type PostTaskJSONRequestBody = Task

// PutTaskJSONRequestBody defines body for PutTask for application/json ContentType.
type PutTaskJSONRequestBody = Task